config/requirements.txt
Django>=5.2.6
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-filter>=25.1
openpyxl>=3.1.2              # For Excel export functionality
pandas>=2.0.3                # Optional, for data handling in marksheet export
xlrd>=2.0.1                  # Optional, if you plan to read Excel files